Break down  Rails::Generator::TestCase into modules
